@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.solution.tecno.androidanimations"
-        minSdkVersion 15
-        targetSdkVersion 27
-        versionCode 34
-        versionName "2.9.1"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        minSdkVersion 19
+        targetSdkVersion 29
+        versionCode 37
+        versionName "3.0.2"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -20,40 +20,35 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
     //Extras
-    implementation 'com.android.volley:volley:1.1.0'
+    implementation 'com.android.volley:volley:1.1.1'
     implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1'
     //edit text material design
     implementation 'com.rengwuxian.materialedittext:library:2.1.4'
     //progress button
     implementation 'com.alirezaahmadi:progress-button:1.1.2'
-    //input dialog
-    implementation 'com.github.javiersantos:MaterialStyledDialogs:2.1'
     //Firebase Notification
-    implementation 'com.google.firebase:firebase-messaging:15.0.2'
+    implementation 'com.google.firebase:firebase-messaging:20.2.1'
     //Circle Image
     implementation 'de.hdodenhof:circleimageview:2.1.0'
-    //SWEET Dialog
-    implementation 'com.github.blennerSilva:AwesomeDialog:2.0.5'
     //Card vie
-    implementation 'com.android.support:cardview-v7:27.1.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
     //SWIPE button
     implementation 'in.shadowfax:proswipebutton:1.1'
     //FAVORITE BUTTON
     implementation 'com.github.ivbaranov:materialfavoritebutton:0.1.5'
-    //CLOUDINARY
-    implementation group: 'com.cloudinary', name: 'cloudinary-android', version: '1.24.0'
     //PICASSO
     implementation 'com.squareup.picasso:picasso:2.71828'
-
+    //GSON
+    implementation 'com.google.code.gson:gson:2.8.6'
 }
 
 apply plugin: 'com.google.gms.google-services'
