@@ -1,29 +1,25 @@
-package com.solution.tecno.androidanimations;
+package com.solution.tecno.androidanimations.adapters;
 
+import android.app.AlertDialog;
 import android.content.ClipboardManager;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.afollestad.materialdialogs.DialogAction;
-import com.afollestad.materialdialogs.MaterialDialog;
 import com.android.volley.DefaultRetryPolicy;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
@@ -31,14 +27,13 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.awesomedialog.blennersilva.awesomedialoglibrary.AwesomeErrorDialog;
-import com.awesomedialog.blennersilva.awesomedialoglibrary.AwesomeProgressDialog;
-import com.awesomedialog.blennersilva.awesomedialoglibrary.AwesomeSuccessDialog;
-import com.awesomedialog.blennersilva.awesomedialoglibrary.AwesomeWarningDialog;
-import com.awesomedialog.blennersilva.awesomedialoglibrary.interfaces.Closure;
 import com.github.ivbaranov.mfb.MaterialFavoriteButton;
-import com.github.javiersantos.materialstyleddialogs.MaterialStyledDialog;
-import com.github.javiersantos.materialstyleddialogs.enums.Style;
+import com.solution.tecno.androidanimations.fragments.AccountsFragment;
+import com.solution.tecno.androidanimations.activities.FirstActivity;
+import com.solution.tecno.androidanimations.R;
+import com.solution.tecno.androidanimations.utils.Credentials;
+import com.solution.tecno.androidanimations.utils.Utils;
+import com.solution.tecno.androidanimations.utils.ViewDialog;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
@@ -51,11 +46,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
 
     List<JSONObject> l;
     Context ctx;
+    ViewDialog viewDialog;
 
-    AwesomeProgressDialog apd;
-    AwesomeSuccessDialog asd;
-    AwesomeErrorDialog aed;
-    AwesomeWarningDialog awd;
     private String base_url="https://www.jadconsultores.com.pe/php_connection/app/bancos_resumen/";
     private EditText diag_et_bank,diag_et_number,diag_et_user_name,diag_et_cci;
     private String bank_edit,account_edit,titular_edit,cci_edit;
@@ -68,37 +60,7 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         ctx=parent.getContext();
         cred = new Credentials(ctx);
-        //create progress dialog
-        apd=new AwesomeProgressDialog(ctx)
-                .setTitle(R.string.app_name)
-                .setMessage("Cargando")
-                .setColoredCircle(R.color.dialogInfoBackgroundColor)
-                .setDialogIconAndColor(R.drawable.ic_dialog_info, R.color.white)
-                .setCancelable(false);
-
-        //create success dialog
-        asd=new AwesomeSuccessDialog(ctx)
-                .setTitle(R.string.app_name)
-                .setMessage("Listo!")
-                .setColoredCircle(R.color.dialogSuccessBackgroundColor)
-                .setDialogIconAndColor(R.drawable.ic_dialog_info, R.color.white)
-                .setCancelable(false);
-
-        //create error dialog
-        aed=new AwesomeErrorDialog(ctx)
-                .setTitle(R.string.app_name)
-                .setMessage("Ocurrió un error")
-                .setColoredCircle(R.color.dialogErrorBackgroundColor)
-                .setDialogIconAndColor(R.drawable.ic_dialog_error,R.color.white)
-                .setCancelable(false);
-
-        awd=new AwesomeWarningDialog(ctx)
-                .setTitle(R.string.app_name)
-                .setMessage("Actualiza tu número de celular para disfrutar de todas las funciones")
-                .setColoredCircle(R.color.dialogWarningBackgroundColor)
-                .setDialogIconAndColor(R.drawable.ic_dialog_warning,R.color.white)
-                .setCancelable(false);
-
+        viewDialog = new ViewDialog((FirstActivity)ctx);
         user_id = cred.getUserId();
 
         return new AccountAdapter.ViewHolder(LayoutInflater.from(ctx).inflate(R.layout.item_account,parent,false));
@@ -117,9 +79,7 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
         final String cci = obj.get("cci").toString().trim();
         final String id = obj.get("id").toString();
         final String fav = obj.get("fav").toString();
-        Random rnd = new Random();
-        int currentColor = Color.argb(255, rnd.nextInt(200), rnd.nextInt(200), rnd.nextInt(200));
-        holder.cardView.setCardBackgroundColor(currentColor);
+
         holder.titular.setText(titular);
         holder.banco.setText(bco);
         holder.cta.setText(cta);
@@ -134,22 +94,13 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
             @Override
             public void onClick(View v) {
                 if(cred.getNetworkStatus().equals("1")){
-                    apd.setMessage("Actualizando...");
-                    apd.show();
                     if(fav.equals("0")){
                         updateFavorite(id,1);
                     }else{
                         updateFavorite(id,0);
                     }
                 }else{
-                    aed.setMessage("Red no disponible");
-                    aed.show();
-                    Handler handler = new Handler();
-                    handler.postDelayed(new Runnable() {
-                        public void run() {
-                            aed.hide();
-                        }
-                    }, 1500);
+                    new Utils().createAlert(ctx,"Red no disponible",1);
                 }
             }
         });
@@ -204,25 +155,26 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
             @Override
             public void onClick(View view) {
                 if(cred.getNetworkStatus().equals("1")){
-                    awd.setButtonText("Eliminar");
-                    awd.setCancelable(true);
-                    awd.setMessage("¿Estás seguro de eliminar esta cuenta?");
-                    awd.setWarningButtonClick(new Closure() {
-                        @Override
-                        public void exec() {
-                            deleteAccount(id);
-                        }
-                    });
-                    awd.show();
+                    viewDialog.hideDialog(0);
+                    new AlertDialog.Builder(ctx)
+                            .setTitle("Eliminar")
+                            .setMessage("¿Estás seguro de eliminar esta cuenta?")
+                            .setPositiveButton("Eliminar", new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    deleteAccount(id);
+                                }
+                            })
+                            .setNegativeButton("Cancelar", new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    dialog.dismiss();
+                                }
+                            })
+                            .show();
                 }else{
-                    aed.setMessage("Red no disponible");
-                    aed.show();
-                    Handler handler = new Handler();
-                    handler.postDelayed(new Runnable() {
-                        public void run() {
-                            aed.hide();
-                        }
-                    }, 1500);
+                    viewDialog.hideDialog(0);
+                    new Utils().createAlert(ctx,"Red no disponible",1);
                 }
             }
         });
@@ -231,22 +183,14 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
             @Override
             public void onClick(View v) {
                 if(cred.getNetworkStatus().equals("0")){
-                    aed.setMessage("Red no disponible");
-                    aed.show();
-                    Handler handler = new Handler();
-                    handler.postDelayed(new Runnable() {
-                        public void run() {
-                            aed.hide();
-                        }
-                    }, 1500);
+                    viewDialog.hideDialog(0);
+                    new Utils().createAlert(ctx,"Red no disponible",1);
                 }else{
-                    apd.setMessage("Obteniendo datos...");
-                    apd.show();
-
+                    viewDialog.showDialog();
+                    viewDialog.hideDialog(1.5);
                     new Handler().postDelayed(new Runnable() {
                         @Override
                         public void run() {
-                            //                        getAccountDetail(account_id);
                             final View layout=LayoutInflater.from(ctx).inflate(R.layout.edit_account_view,null);
                             diag_et_bank=layout.findViewById(R.id.diag_et_bank_edit);
                             diag_et_number=layout.findViewById(R.id.diag_et_account_edit);
@@ -258,42 +202,30 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                             diag_et_user_name.setText(titular);
                             diag_et_cci.setText(cci);
 
-                            apd.hide();
-                            asd.show();
-                            new Handler().postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    asd.hide();
-                                    new MaterialStyledDialog.Builder(ctx)
-                                            .setStyle(Style.HEADER_WITH_TITLE)
-                                            .setTitle("Editar cuenta")
-                                            .setDescription("Edita tu cuenta para compartirla rápidamente")
-                                            .setPositiveText("Guardar")
-                                            .onPositive(new MaterialDialog.SingleButtonCallback() {
-                                                @Override
-                                                public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                                                    dialog.dismiss();
-                                                    apd.setMessage("Guardando...");
-                                                    String bank = diag_et_bank.getText().toString();
-                                                    String number = diag_et_number.getText().toString();
-                                                    String user_name =diag_et_user_name.getText().toString();
-                                                    String cci = diag_et_cci.getText().toString();
-                                                    updateAccount(id,bank,number,user_name,cci);
-                                                }
-                                            })
-                                            .setNegativeText("Cancelar")
-                                            .onNegative(new MaterialDialog.SingleButtonCallback() {
-                                                @Override
-                                                public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                                                    dialog.dismiss();
-                                                }
-                                            })
-                                            .setCustomView(layout)
-                                            .show();
-                                }
-                            },1000);//1 sec
+                            new AlertDialog.Builder(ctx)
+                                    .setTitle("Editar cuenta")
+                                    .setMessage("Edita tu cuenta para compartirla rápidamente")
+                                    .setPositiveButton("Guardar", new DialogInterface.OnClickListener() {
+                                        @Override
+                                        public void onClick(DialogInterface dialog, int which) {
+                                            dialog.dismiss();
+                                            String bank = diag_et_bank.getText().toString();
+                                            String number = diag_et_number.getText().toString();
+                                            String user_name =diag_et_user_name.getText().toString();
+                                            String cci = diag_et_cci.getText().toString();
+                                            updateAccount(id,bank,number,user_name,cci);
+                                        }
+                                    })
+                                    .setNegativeButton("Cancelar", new DialogInterface.OnClickListener() {
+                                        @Override
+                                        public void onClick(DialogInterface dialog, int which) {
+                                            dialog.dismiss();
+                                        }
+                                    })
+                                    .setView(layout)
+                                    .show();
                         }
-                    }, 2500);
+                    }, 1500);
                 }
             }
         });
@@ -326,15 +258,13 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
             ic_share_cci=itemView.findViewById(R.id.ic_share_cci);
             ic_copy_cta=itemView.findViewById(R.id.ic_copy_cta);
             ic_copy_cci=itemView.findViewById(R.id.ic_copy_cci);
-//            ic_edit_card=itemView.findViewById(R.id.ic_edit_card);
             ic_delete_card=itemView.findViewById(R.id.ic_delete_card);
             fav = itemView.findViewById(R.id.item_fav);
         }
     }
 
     private void getAccounts(final String user_id) {
-        apd.setMessage("Cargando...");
-        apd.show();
+        viewDialog.showDialog();
         RequestQueue queue = Volley.newRequestQueue(ctx);
         String params="?user_id="+Integer.parseInt(user_id);
         String url = base_url+"getUserAccounts.php"+params;
@@ -346,33 +276,24 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                         try {
                             JSONArray ja=(JSONArray)jp.parse(response);
                             l.clear();
+                            notifyDataSetChanged();
                             for(int i=0;i<ja.size();i++){
                                 JSONObject item=(JSONObject)ja.get(i);
                                 l.add(item);
                             }
                             notifyDataSetChanged();
-                            apd.hide();
-                            asd.show();
+                            viewDialog.hideDialog(0);
                             //wait 3 seconds to hide success dialog
-                            Handler handler = new Handler();
-                            handler.postDelayed(new Runnable() {
+                            new Handler().postDelayed(new Runnable() {
                                 public void run() {
-                                    asd.hide();
                                     verifiedPhoneNumber();
                                 }
                             }, 1500);   //3 seconds
+
                         } catch (Exception e) {
                             cred.registerError(e.toString(),user_id);
-                            apd.hide();
-                            aed.setMessage("No se pudo obtener la información");
-                            aed.show();
-                            //wait 3 seconds to hide success dialog
-                            Handler handler = new Handler();
-                            handler.postDelayed(new Runnable() {
-                                public void run() {
-                                    aed.hide();
-                                }
-                            }, 3000);
+                            viewDialog.hideDialog(0);
+                            new Utils().createAlert(ctx,"No se pudo obtener la información",1);
                         }
                     }
                 },
@@ -380,16 +301,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     @Override
                     public void onErrorResponse(VolleyError error) {
                         cred.registerError(error.toString(),user_id);
-                        apd.hide();
-                        aed.setMessage("No se pudo obtener la información");
-                        aed.show();
-                        //wait 3 seconds to hide success dialog
-                        Handler handler = new Handler();
-                        handler.postDelayed(new Runnable() {
-                            public void run() {
-                                aed.hide();
-                            }
-                        }, 3000);
+                        viewDialog.hideDialog(0);
+                        new Utils().createAlert(ctx,"No se pudo obtener la información",1);
                         // error
                         Log.d("Error.Response", error.toString());
                     }
@@ -402,8 +315,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
         queue.add(postRequest);
     }
 
-    private void updateAccount(String id,String bank, String number, String user_name,String cci) {
-
+    private void updateAccount(String id,String bank, String number, String user_name,String cci)
+    {
         RequestQueue queue = Volley.newRequestQueue(ctx);
         String params="?id="+id+
                 "&bank="+Uri.encode(bank)+
@@ -416,19 +329,10 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     @Override
                     public void onResponse(String response) {
                         try {
-                            apd.hide();
                             getAccounts(user_id);
                         } catch (Exception e) {
                             cred.registerError(e.getMessage(),user_id);
-                            apd.hide();
-                            aed.setMessage("Error al actualizar");
-                            aed.show();
-                            new Handler().postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    aed.hide();
-                                }
-                            }, 2500);
+                            new Utils().createAlert(ctx,"Error al actualizar",1);
                         }
                     }
                 },
@@ -437,15 +341,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     public void onErrorResponse(VolleyError error) {
                         // error
                         cred.registerError(error.getMessage(),user_id);
-                        apd.hide();
-                        aed.setMessage("Error al actualizar");
-                        aed.show();
-                        new Handler().postDelayed(new Runnable() {
-                            @Override
-                            public void run() {
-                                aed.hide();
-                            }
-                        }, 2500);
+                        viewDialog.hideDialog(0);
+                        new Utils().createAlert(ctx,"Error al actualizar",1);
                     }
                 }
         );
@@ -457,29 +354,20 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
     }
 
     private void updateFavorite(String id,int fav) {
-
         RequestQueue queue = Volley.newRequestQueue(ctx);
-        String params="?id="+id+
-                "&fav="+fav;
+        String params="?id="+id+"&fav="+fav;
         String url = base_url+"updateFavorite.php"+params;
         StringRequest postRequest = new StringRequest(Request.Method.POST, url,
                 new Response.Listener<String>() {
                     @Override
                     public void onResponse(String response) {
                         try {
-                            apd.hide();
+                            viewDialog.hideDialog(0);
                             getAccounts(user_id);
                         } catch (Exception e) {
                             cred.registerError(e.getMessage(),user_id);
-                            apd.hide();
-                            aed.setMessage("Error al actualizar");
-                            aed.show();
-                            new Handler().postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    aed.hide();
-                                }
-                            }, 2500);
+                            viewDialog.hideDialog(0);
+                            new Utils().createAlert(ctx,"Error al actualizar",1);
                         }
                     }
                 },
@@ -488,15 +376,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     public void onErrorResponse(VolleyError error) {
                         // error
                         cred.registerError(error.getMessage(),user_id);
-                        apd.hide();
-                        aed.setMessage("Error al actualizar");
-                        aed.show();
-                        new Handler().postDelayed(new Runnable() {
-                            @Override
-                            public void run() {
-                                aed.hide();
-                            }
-                        }, 2500);
+                        viewDialog.hideDialog(0);
+                        new Utils().createAlert(ctx,"Error al actualizar",1);
                     }
                 }
         );
@@ -508,9 +389,6 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
     }
 
     public void deleteAccount(String id) {
-        apd.setMessage("Eliminando...");
-        apd.show();
-
         RequestQueue queue = Volley.newRequestQueue(ctx);
         String params="?id="+id;
         String url = base_url+"deleteAccount.php"+params;
@@ -520,19 +398,12 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     @Override
                     public void onResponse(String response) {
                         try {
-                            apd.hide();
+                            viewDialog.hideDialog(0);
                             getAccounts(user_id);
                         } catch (Exception e) {
                             cred.registerError(e.getMessage(),user_id);
-                            apd.hide();
-                            aed.setMessage("Ocurrió un error al eliminar");
-                            aed.show();
-                            new Handler().postDelayed(new Runnable() {
-                                @Override
-                                public void run() {
-                                    aed.hide();
-                                }
-                            }, 2500);
+                            viewDialog.hideDialog(0);
+                            new Utils().createAlert(ctx,"Ocurrió un error al eliminar",1);
                         }
                     }
                 },
@@ -541,15 +412,8 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
                     public void onErrorResponse(VolleyError error) {
                         // error
                         cred.registerError(error.getMessage(),user_id);
-                        apd.hide();
-                        aed.setMessage("Ocurrió un error al eliminar");
-                        aed.show();
-                        new Handler().postDelayed(new Runnable() {
-                            @Override
-                            public void run() {
-                                aed.hide();
-                            }
-                        }, 2500);
+                        viewDialog.hideDialog(0);
+                        new Utils().createAlert(ctx,"Ocurrió un error al eliminar",1);
                     }
                 }
         );
@@ -561,15 +425,11 @@ public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHold
     }
 
     private void verifiedPhoneNumber(){
-        asd.hide();
+        viewDialog.hideDialog(0);
         if(cred.getPhoneNumber().equals("0")){
-            awd.show();
-            new Handler().postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    awd.hide();
-                }
-            },3500);
+            new Utils().createAlert(ctx,"Actualiza tu número de celular para disfrutar de todas las funciones",1);
+        }else{
+            viewDialog.hideDialog(0);
         }
     }
 }
